rgw: generate config in init container
