+++ pkg/operator/ceph/object/rgw_test.go
@@ -17,20 +17,16 @@ limitations under the License.
 package object
 
 import (
-	"fmt"
 	"io/ioutil"
 	"os"
 	"testing"
 
 	cephv1beta1 "github.com/rook/rook/pkg/apis/ceph.rook.io/v1beta1"
 	"github.com/rook/rook/pkg/clusterd"
-	"github.com/rook/rook/pkg/operator/k8sutil"
 	testop "github.com/rook/rook/pkg/operator/test"
 	exectest "github.com/rook/rook/pkg/util/exec/test"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/fake"
 )
@@ -93,72 +89,6 @@ func validateStart(t *testing.T, store cephv1beta1.ObjectStore, clientset *fake.
 	assert.Equal(t, 1, len(secret.StringData))
 }
 
-func TestPodSpecs(t *testing.T) {
-	store := simpleStore()
-	store.Spec.Gateway.Resources = v1.ResourceRequirements{
-		Limits: v1.ResourceList{
-			v1.ResourceCPU: *resource.NewQuantity(100.0, resource.BinarySI),
-		},
-		Requests: v1.ResourceList{
-			v1.ResourceMemory: *resource.NewQuantity(1337.0, resource.BinarySI),
-		},
-	}
-
-	s := makeRGWPodSpec(store, "rook/rook:myversion", true)
-	assert.NotNil(t, s)
-	//assert.Equal(t, instanceName(store), s.Name)
-	assert.Equal(t, v1.RestartPolicyAlways, s.Spec.RestartPolicy)
-	assert.Equal(t, 2, len(s.Spec.Volumes))
-	assert.Equal(t, "rook-data", s.Spec.Volumes[0].Name)
-	assert.Equal(t, k8sutil.ConfigOverrideName, s.Spec.Volumes[1].Name)
-
-	assert.Equal(t, instanceName(store), s.ObjectMeta.Name)
-	assert.Equal(t, appName, s.ObjectMeta.Labels["app"])
-	assert.Equal(t, store.Namespace, s.ObjectMeta.Labels["rook_cluster"])
-	assert.Equal(t, store.Name, s.ObjectMeta.Labels["rook_object_store"])
-	assert.Equal(t, 0, len(s.ObjectMeta.Annotations))
-
-	cont := s.Spec.Containers[0]
-	assert.Equal(t, "rook/rook:myversion", cont.Image)
-	assert.Equal(t, 2, len(cont.VolumeMounts))
-
-	assert.Equal(t, 6, len(cont.Args))
-	assert.Equal(t, "ceph", cont.Args[0])
-	assert.Equal(t, "rgw", cont.Args[1])
-	assert.Equal(t, "--config-dir=/var/lib/rook", cont.Args[2])
-	assert.Equal(t, fmt.Sprintf("--rgw-name=%s", "default"), cont.Args[3])
-	assert.Equal(t, fmt.Sprintf("--rgw-port=%d", 123), cont.Args[4])
-	assert.Equal(t, fmt.Sprintf("--rgw-secure-port=%d", 0), cont.Args[5])
-
-	assert.Equal(t, (7 + len(k8sutil.ClusterDaemonEnvVars())), len(cont.Env))
-
-	assert.Equal(t, "100", cont.Resources.Limits.Cpu().String())
-	assert.Equal(t, "1337", cont.Resources.Requests.Memory().String())
-}
-
-func TestSSLPodSpec(t *testing.T) {
-	store := simpleStore()
-	store.Spec.Gateway.SSLCertificateRef = "mycert"
-	store.Spec.Gateway.SecurePort = 443
-
-	s := makeRGWPodSpec(store, "v1.0", true)
-	assert.NotNil(t, s)
-	assert.Equal(t, instanceName(store), s.Name)
-	assert.Equal(t, 3, len(s.Spec.Volumes))
-	assert.Equal(t, certVolumeName, s.Spec.Volumes[2].Name)
-	assert.True(t, s.Spec.HostNetwork)
-	assert.Equal(t, v1.DNSClusterFirstWithHostNet, s.Spec.DNSPolicy)
-
-	cont := s.Spec.Containers[0]
-	assert.Equal(t, 3, len(cont.VolumeMounts))
-	assert.Equal(t, certVolumeName, cont.VolumeMounts[2].Name)
-	assert.Equal(t, certMountPath, cont.VolumeMounts[2].MountPath)
-
-	assert.Equal(t, 7, len(cont.Args))
-	assert.Equal(t, fmt.Sprintf("--rgw-secure-port=%d", 443), cont.Args[5])
-	assert.Equal(t, fmt.Sprintf("--rgw-cert=%s/%s", certMountPath, certFilename), cont.Args[6])
-}
-
 func TestCreateObjectStore(t *testing.T) {
 	executor := &exectest.MockExecutor{
 		MockExecuteCommandWithCombinedOutput: func(debug bool, actionName, command string, args ...string) (string, error) {
@@ -184,39 +114,6 @@ func TestCreateObjectStore(t *testing.T) {
 
 	// create the pools
 	err := CreateStore(context, store, "1.2.3.4", false, []metav1.OwnerReference{})
-	assert.Nil(t, err)
-}
-
-func TestValidateSpec(t *testing.T) {
-	context := &clusterd.Context{Executor: &exectest.MockExecutor{}}
-
-	// valid store
-	s := simpleStore()
-	err := validateStore(context, s)
-	assert.Nil(t, err)
-
-	// no name
-	s.Name = ""
-	err = validateStore(context, s)
-	assert.NotNil(t, err)
-	s.Name = "default"
-	err = validateStore(context, s)
-	assert.Nil(t, err)
-
-	// no namespace
-	s.Namespace = ""
-	err = validateStore(context, s)
-	assert.NotNil(t, err)
-	s.Namespace = "mycluster"
-	err = validateStore(context, s)
-	assert.Nil(t, err)
-
-	// no replication or EC
-	s.Spec.MetadataPool.Replicated.Size = 0
-	err = validateStore(context, s)
-	assert.NotNil(t, err)
-	s.Spec.MetadataPool.Replicated.Size = 1
-	err = validateStore(context, s)
 	assert.Nil(t, err)
 }
